Add files via upload
The following files are:

AdminMainScreenController: Controls are similar to what Employees would see on their end, just slightly different due to their position.

AnnouncementsController: This helps pull the announcement information made by the Admin from the MySQL database, and displays the announcements for Employees.

Database Connection: This file is what actually connects the MySQL database to the application itself.

UserAuthentication: This file helps verify Users based on their position in a company, whether they are employee or admin. When a User logs in, they will be directed to their proper screen. The MySQL database created has two separate User tables, one for Employees, and the other for Admin/s.