Add files via upload
The following files are:

AdminTimeDisplayController: This file is just to separate from the TimeDisplayController (Employees). This was created to prevent Admins from swapping from one screen to the EmployeeMainScreen. For example, an Admin would log in, click Time Display or Send Announcements, and when they click to go back to their "Main Screen", they are brought back to the EmployeeMainScreen instead of AdminMainScreen.

Announcement: This was created to work with the TableView and ObservableList, to grab and store Announcement information and set the manually inputed information from the Admin in the TableView that is added for the Employees to see. 

CreateAnnouncementsController: This was created to take Admin's inputs, so Date Assigned, Message, and Sender, and store that information in the MySQL database table "Announcements". That information is then pulled from the AnnouncementsController and displayed for Employees when they click on the Announcements button on their end.